Clone blog post loading logic into LatestPosts
Due to missing information when using `Astro.glob` to load blog posts,
we need to a way to load blog posts and have the correct trailing url
path for linking to the blog post. Unfortunately, `Astro.glob` does not
provide this information, so we need to clone the logic for loading blog
posts into the `LatestPosts` component.